Only allows access from cloudfront
There are a list of published IP address ranges which AWS uses for
cloudfront. These can be added to security groups so that you can only
access jenkins through cloudfront and https and not directly through the
network load balancer.

There will be a lambda which updates the address ranges which will come
later.